Make LogStream into an interface
so users of library can write tests that use it.
