+++ src/main/java/com/spotify/docker/client/LogStream.java
@@ -17,71 +17,14 @@
 
 package com.spotify.docker.client;
 
-import static com.google.common.base.Charsets.UTF_8;
-
-import com.google.common.base.Throwables;
-import com.google.common.collect.AbstractIterator;
-import com.google.common.io.Closer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.Closeable;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.ByteBuffer;
+import java.util.Iterator;
 
-public class LogStream extends AbstractIterator<LogMessage> implements Closeable {
+public interface LogStream extends Iterator<LogMessage>, Closeable {
 
-  private static final Logger log = LoggerFactory.getLogger(LogStream.class);
-
-  private final LogReader reader;
-  private volatile boolean closed;
-
-  LogStream(final InputStream stream) {
-    this.reader = new LogReader(stream);
-  }
-
-  @Override
-  protected void finalize() throws Throwable {
-    super.finalize();
-    if (!closed) {
-      log.warn(this + " not closed properly");
-      close();
-    }
-  }
-
-  @Override
-  protected LogMessage computeNext() {
-    final LogMessage message;
-    try {
-      message = reader.nextMessage();
-    } catch (IOException e) {
-      throw Throwables.propagate(e);
-    }
-    if (message == null) {
-      return endOfData();
-    }
-    return message;
-  }
-
-  @Override
-  public void close() {
-    closed = true;
-    try {
-      reader.close();
-    } catch (IOException e) {
-      throw Throwables.propagate(e);
-    }
-  }
-
-  public String readFully() {
-    final StringBuilder stringBuilder = new StringBuilder();
-    while (hasNext()) {
-      stringBuilder.append(UTF_8.decode(next().content()));
-    }
-    return stringBuilder.toString();
-  }
+  String readFully();
 
   /**
    * Attaches two {@link java.io.OutputStream}s to the {@link LogStream}.  Closes the streams after
@@ -92,9 +35,7 @@ public class LogStream extends AbstractIterator<LogMessage> implements Closeable
    * @throws IOException if an I/O error occurs
    * @see #attach(OutputStream, OutputStream, boolean) for control over stream lifecycles
    */
-  public void attach(final OutputStream stdout, final OutputStream stderr) throws IOException {
-    attach(stdout, stderr, true);
-  }
+  void attach(OutputStream stdout, OutputStream stderr) throws IOException;
 
   /**
    * Attaches two {@link java.io.OutputStream}s to the {@link LogStream}.
@@ -149,40 +90,5 @@ public class LogStream extends AbstractIterator<LogMessage> implements Closeable
    * @see java.io.PipedInputStream
    * @see java.io.PipedOutputStream
    */
-  public void attach(final OutputStream stdout, final OutputStream stderr, boolean closeAtEOF)
-      throws IOException {
-    final Closer closer = Closer.create();
-    try {
-      if (closeAtEOF) {
-        closer.register(stdout);
-        closer.register(stderr);
-      }
-
-      while (this.hasNext()) {
-        final LogMessage message = this.next();
-        final ByteBuffer content = message.content();
-
-        assert content.hasArray();
-
-        switch (message.stream()) {
-          case STDOUT:
-            stdout.write(content.array(), content.position(), content.remaining());
-            stdout.flush();
-            break;
-          case STDERR:
-            stderr.write(content.array(), content.position(), content.remaining());
-            stderr.flush();
-            break;
-          case STDIN:
-          default:
-            break;
-        }
-      }
-    } catch (Throwable t) {
-      throw closer.rethrow(t);
-    } finally {
-      closer.close();
-    }
-  }
-
+  void attach(OutputStream stdout, OutputStream stderr, boolean closeAtEOF) throws IOException;
 }
